requirements.txt
requests
pandas
numpy
psycopg2-binary
python-dotenv
openpyxl
xlrd
sqlalchemy
